refactor(v0.5): flatten agent/modes/separated/ into agent/modes/
The separated/ package held a single file (composer.py). The "separated"
MODES key is unchanged (wire-visible mode id).

SPEC: 20260730-1659_v0.5-breaking-refactor-architecture-cleanup
Phase: 6/7 (internal cleanup)
